+++ app/containers/Network/containers/AccessPoints/index.module.scss
@@ -1,0 +1,5 @@
+.tabColumn {
+  display: flex;
+  flex-direction: column;
+  padding: 0 30px;
+}
