+++ app/containers/Network/containers/ClientDeviceDetails/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { useParams } from 'react-router-dom';
+
+const ClientDeviceDetails = () => {
+  const { id } = useParams();
+  return <h1>Client Device Details: {id}</h1>;
+};
+
+export default ClientDeviceDetails;
