--- app/containers/Network/index.module.scss
@@ -1,11 +1,0 @@
-.spinner {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-.tabColumn {
-  display: flex;
-  flex-direction: column;
-  padding: 0 30px;
-}
